Initial working commit
As of this commit this extension creates a new menu-item which always
opens the right-clicked twitter image in a new tab.

This works for twitter images in and out of the gallery view (the view
where right-click normally will not allow opening an image).

In addition, the image is always specifically the ":orig" copy of the
image.