Add support for opening the mp4 source of a video
Since the mp4 lives in an iframe, we also now need to inject the content
script into all iframes.

This isn't really the 'original file', but it's the best we can get from
twitter as far as I know.

This is still in the spirit of the extension because the right-click
menu for those cards doesn't have a 'view video' option due to the heavy
use of divs.